Extract integration test base class to its own file
